requirements.txt
flask
flask-sqlalchemy
sqlalchemy-migrate
flask-wtf
flask-openid
flask-login